// src/app/_metronic/shared/shared/models/options.ts
export interface Options {
    value: any; // string | number;
    label: string;
    //==============
    isActive?: boolean;
    phone?: string;
    type?: string;
    role?: string;
    dvt?: string;
    giaThamKhao?: string;
    startDate?: string;
    endDate?: string;
    meta?: any; // all
}